Added Note for sitemetadata title in Head export
It may have been just me, but when I was following this part of the tutorial (https://www.gatsbyjs.com/docs/tutorial/part-4/#:~:text=export%20default%20BlogPage-,Pro%20Tip%3A,-The%20Gatsby%20Head). I copied and pasted the code and replaced what I already had in the Head of my blog page, since I was following it sequentially. I faced various error since the GraphQL query didn't have the data for the metadata; only to realize this was already handled by the Seo components some steps ago. So I wanted to make a note to make it clear that this wasn't necessary since we had the Seo component already.